rendering a 302 for renamed assets
