+++ packages/cli/src/actions/Builder.ts
@@ -1,5 +1,4 @@
 import {
-  BuildFailedError,
   FabBuildStep,
   FabConfig,
   FabPluginRuntime,
@@ -11,6 +10,7 @@ import {
   s_sume,
 } from '@fab/core'
 import { Compiler } from './Compiler'
+import { Generator } from './Generator'
 import { relativeToConfig } from '../helpers/paths'
 
 export default class Builder {
@@ -86,15 +86,7 @@ export default class Builder {
       await builder(plugin_args, proto_fab)
     }
 
-    // After build, there should only be files in the expected places (server.js, _assets)
-    const invalid_reason = proto_fab.errorsPreventingCompilation()
-    if (invalid_reason) {
-      throw new BuildFailedError(`FAB is not ready for compilation.
-${invalid_reason}
-You might need to add @fab/rewire-assets to your 'build' config. See https://fab.dev/packages/rewire-assets for more information about what this module is and why it's needed.
-`)
-    }
-
     await Compiler.compile(proto_fab, runtime_plugins)
+    await Generator.generate(proto_fab)
   }
 }
